+++ inst/app/tools/app/help.R
@@ -18,7 +18,7 @@ output$help_text <- renderUI({
 #######################################
 ## Main function of help menu
 #######################################
-append_help <- function(help_str, help_path, Rmd = TRUE) {
+append_help <- function(help_str, help_path, lic = "nc", Rmd = TRUE) {
   if (length(input[[help_str]]) == 0) return()
   help_block <- get(help_str)
   local_hd <- help_block[which(help_block %in% input[[help_str]])]
@@ -32,11 +32,16 @@ append_help <- function(help_str, help_path, Rmd = TRUE) {
   }
   mathjax_script <- ifelse(Rmd, "<script>if (window.MathJax) MathJax.Hub.Typeset();</script>", "")
 
-  cc <- getOption("radiant.help.cc", default = "")
+  if (lic == "nc") {
+    cc <- getOption("radiant.help.cc", default = "")
+  } else {
+    cc <- getOption("radiant.help.cc.sa", default = "")
+  }
 
   ## remove ` from report.md
   paste(
-    gsub("(\"> )`", "\\1", all_help) %>% gsub("`( </td>)", "\\1", .),
+    gsub("(\"> )`", "\\1", all_help) %>% 
+    gsub("`( </td>)", "\\1", .),
     "\n", mathjax_script, "\n", cc
   ) %>% HTML()
 }
@@ -59,7 +64,9 @@ help_data <- c(
   "Pivot" = "pivotr.md", "Explore" = "explore.md", "Transform" = "transform.md",
   "Combine" = "combine.md", "Report > Rmd" = "report_rmd.md", "Report > R" = "report_r.md"
 )
-output$help_data <- reactive(append_help("help_data", file.path(getOption("radiant.path.data"), "app/tools/help/")))
+output$help_data <- reactive(
+  append_help("help_data", file.path(getOption("radiant.path.data"), "app/tools/help/"))
+)
 
 observeEvent(input$help_data_all, {
   help_switch(input$help_data_all, "help_data")
@@ -70,8 +77,10 @@ observeEvent(input$help_data_none, {
 
 help_data_panel <-
   wellPanel(
-    HTML("<label>Data menu: <i id='help_data_all' title='Check all' href='#' class='action-button glyphicon glyphicon-ok'></i>
-              <i id='help_data_none' title='Uncheck all' href='#' class='action-button glyphicon glyphicon-remove'></i></label>"),
+    HTML(
+      "<label>Data menu: <i id='help_data_all' title='Check all' href='#' class='action-button glyphicon glyphicon-ok'></i>
+       <i id='help_data_none' title='Uncheck all' href='#' class='action-button glyphicon glyphicon-remove'></i></label>"
+    ),
     checkboxGroupInput("help_data", NULL, help_data, selected = state_group("help_data"), inline = TRUE)
   )
 
@@ -113,15 +122,6 @@ observeEvent(input$help_keyboard, {
              td("Blue button load"), td("CMD-o"), td("CTRL-o"))
         )
       ),
-      # h4("Report"),
-      # withTags(
-      #   table(style = "width: 80%; margin-left: 10%;",
-      #     tr(class = "border_bottom",
-      #        td(b("Function")), td(b("Mac")), td(b("Windows & Linux"))),
-      #     tr(class = "padding_top",
-      #       td("Read files"), td("Shift-CMD-o"), td("Shift-CTRL-o"))
-      #   )
-      # ),
       footer = modalButton("OK"),
       size = "s",
       easyClose = TRUE